refactor(mcp): F-4 wave 3j-split — split deps.go into per-tool files
Pure file-level split. Zero behavior change, zero interface change —
only moves code between files so each tool body lives next to its
matching test file.

deps.go went from 1740 LOC (interface + 21 tool bodies + helpers +
constants) to 131 LOC (just the Deps interface and SearchResult
type). Tool bodies move to 8 new files, one per theme:

  tool_data.go              (266 LOC) Delete, Prune, ListData, Add
  tool_feedback.go          (148 LOC) AddFeedback, GetFeedbackStats,
                                      SetContext
  tool_memory.go            (315 LOC) ListMemories, PinMemory,
                                      UnpinMemory, WakeUp + helpers
                                      (extractOwnerID lives here; used
                                      by save/recall too)
  tool_chat.go              (194 LOC) SaveChat, RecallChat, SearchChats
  tool_diary.go             (140 LOC) DiaryWrite, DiaryRead
  tool_query_entity.go      (172 LOC) QueryEntity + helpers
  tool_save_recall_memory.go (286 LOC) SaveMemory, RecallMemory + helpers
  tool_cognify.go           (199 LOC) Cognify, CognifyStatus,
                                      runCognifyPipeline

Each file carries only the imports its bodies need — e.g. tool_diary.go
doesn't pull pkg/orchestrator, tool_feedback.go doesn't pull
database/sql, tool_cognify.go doesn't pull strings. Naming matches the
existing test files (tool_memory_test.go, tool_chat_diary_test.go, etc.)
so jump-to-definition works cleanly both directions.

Why now: 3j pushed deps.go past 1700 LOC. Next wave is toolSearch
(~500 LOC) which would push it to ~2200. Splitting before 3k keeps
the PR diff focused on the tool migration, not "this file is too long".

go vet ./... clean. go test -race -count=1 ./... = 35 PASS / 0 FAIL.
